requirements.txt
certifi==2018.4.16
click==6.7
dulwich==0.19.5
PyYAML==3.13
six==1.11.0
urllib3==1.23
